# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Catalogo de vinos",
        "description": "Descripción de mi API",
        "termsOfService": "https://example.com/terms/",
        "contact": {
            "email": "jesquuliche@hotmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/denominacion": {
            "get": {
                "tags": [
                    "Denominacion"
                ],
                "summary": "Get list of denominaciones",
                "description": "Returns list of denominaciones",
                "operationId": "getDenominacionList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "nombre": {
                                                "type": "string",
                                                "example": "Denominacion 1"
                                            },
                                            "descripcion": {
                                                "type": "string",
                                                "example": "Descripción de la denominacion 1"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Denominacion"
                ],
                "summary": "Store a new denominacion",
                "description": "Stores a new denominacion and returns it",
                "operationId": "storeDenominacion",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Nueva Denominacion"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Descripción de la nueva denominacion"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "nombre": {
                                            "type": "string",
                                            "example": "Nueva Denominacion"
                                        },
                                        "descripcion": {
                                            "type": "string",
                                            "example": "Descripción de la nueva denominacion"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/denominacion/{id}": {
            "get": {
                "tags": [
                    "Denominacion"
                ],
                "summary": "Get denominacion by ID",
                "description": "Returns a single denominacion",
                "operationId": "getDenominacionById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of denominacion to return",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "nombre": {
                                            "type": "string",
                                            "example": "Denominacion 1"
                                        },
                                        "descripcion": {
                                            "type": "string",
                                            "example": "Descripción de la denominacion 1"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Denominacion not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Denominacion"
                ],
                "summary": "Update an existing denominacion",
                "description": "Updates an existing denominacion and returns it",
                "operationId": "updateDenominacion",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of denominacion to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Denominacion Actualizada"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Descripción de la denominacion actualizada"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "nombre": {
                                            "type": "string",
                                            "example": "Denominacion Actualizada"
                                        },
                                        "descripcion": {
                                            "type": "string",
                                            "example": "Descripción de la denominacion actualizada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Denominacion not found"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Denominacion"
                ],
                "summary": "Delete a denominacion",
                "description": "Deletes a single denominacion",
                "operationId": "deleteDenominacion",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of denominacion to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Denominacion not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}